% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_pod.R
\name{parse_data}
\alias{parse_data}
\title{Prepare data for fitting}
\usage{
parse_data(xcol = "dose_uM", groups = "treatment", minD = 7,
  dataset = summary_data)
}
\arguments{
\item{xcol}{string, column name of independant variable (usually the concentration).
Default: "dose_uM"}

\item{groups}{string, column name of grouping variable which entries uniquely identify
    Default: "treatment"
the replicate set of concentration response curves for a certain condition.}

\item{minD}{integer, minimum number of unique concentrations
per concentration response curve, curves with less are removed
    Default: 7}

\item{dataset}{dataframe containing the x, y and groups columns.
Default: summary_data}
}
\description{
Parses data for the \code{pre_data} function.
   Validates if data is a summary data file & creates the tidy format.
   Tidy format is created by checking for a column name \code{variable}.
   Returns class of either type \code{"list" "prepped_time"} or \code{"list" "prepped_data"},
   depending if a \code{timeID} or \code{timeAfterExposure} column was found with multiple unique values.
}
\examples{
prep_data(xcol = "dose_uM", groups = "treatment",
         minD = 7, dataset = test_data)

}
\author{
Steven Wink
}
