% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Typical dose response GFP reporter data}
\format{A data frame with 911 rows and 4 variables:
\describe{
  \item{plateID}{plate id functions as replicate id in this case, 3 unique ids}
  \item{treatment}{set of treatments, with DMSO as negative control}
  \item{dose_uM}{concentrations}
  \item{GFP_int}{GFP intensity levels, fraction of GFP positive cells}
}}
\source{
\url{DDS LACDR LU}
}
\usage{
test_data
}
\description{
Data from a recent publication
}
\keyword{datasets}
