% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_pod.R
\name{model_pod}
\alias{model_pod}
\title{fit piece wise splines}
\usage{
model_pod(xcol, ycol, dataset, respLev, groups)
}
\arguments{
\item{xcol}{string, column name of dependant variable.}

\item{ycol}{string, column name of independent variable/ measure variable.}

\item{dataset}{dataframe containing the x, y and groups columns.}

\item{respLev}{Numeric, A number representing threshold for point of departure.}

\item{groups}{string, column name of grouping variable.}
}
\value{
dataframe with the point of departure estimates
}
\description{
fit piece wise splines
}
\examples{
model_pod(xcol = "GFP_int", ycol = "dose_uM", dataset = test_data,
      respL = 0.1,  groups = "treatment")

}
